Script/IoC: Move some spell scripts from spell_generic to isle_of_conquest cpp
